fixed compilation issues around Int replaced with Long

+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/order/request/OrderItemOptionFileDeleteRequest.kt
@@ -6,14 +6,14 @@ import com.ecwid.apiclient.v3.impl.RequestInfo
 data class OrderItemOptionFileDeleteRequest(
 		val orderNumber: Int = 0,
 		val orderIdentity: String = "",
-		val orderItemId: Int = 0,
+		val orderItemId: Long = 0,
 		val optionName: String = "",
 		val fileId: Int = 0
 ) : ApiRequest {
 
 	constructor(
 			orderNumber: Int = 0,
-			orderItemId: Int = 0,
+			orderItemId: Long = 0,
 			optionName: String = "",
 			fileId: Int = 0
 	) : this(
